chore(geospatial): Convert to .ts
