+++ modules/culling/src/lib/bounding-volumes/axis-aligned-bounding-box.ts
@@ -1,11 +1,38 @@
+import {BoundingVolume} from './bounding-volume';
 import {Vector3} from '@math.gl/core';
+import Plane from '../plane';
 import {INTERSECTION} from '../../constants';
 
 const scratchVector = new Vector3();
 const scratchNormal = new Vector3();
 
-export default class AxisAlignedBoundingBox {
-  constructor(minimum = [0, 0, 0], maximum = [0, 0, 0], center = null) {
+/**
+ * An axis aligned bounding box - aligned with coordinate axes
+ * @see BoundingVolume
+ * @see BoundingRectangle
+ * @see OrientedBoundingBox
+ */
+export default class AxisAlignedBoundingBox implements BoundingVolume {
+  /** The center point of the bounding box. */
+  readonly center: Vector3;
+  /** The positive half diagonal of the bounding box. */
+  readonly halfDiagonal: Vector3;
+  /** The minimum point defining the bounding box. [0, 0, 0] for empty box */
+  readonly minimum: Vector3;
+  /** The maximum point defining the bounding box. [0, 0, 0] for empty box */
+  readonly maximum: Vector3;
+
+  /**
+   * Creates an instance of an AxisAlignedBoundingBox from the minimum and maximum points along the x, y, and z axes.
+   * @param minimum=[0, 0, 0] The minimum point along the x, y, and z axes.
+   * @param maximum=[0, 0, 0] The maximum point along the x, y, and z axes.
+   * @param center The center of the box; automatically computed if not supplied.
+   */
+  constructor(
+    minimum: readonly number[] = [0, 0, 0],
+    maximum: readonly number[] = [0, 0, 0],
+    center?: readonly number[]
+  ) {
     // If center was not defined, compute it.
     center = center || scratchVector.copy(minimum).add(maximum).scale(0.5);
     this.center = new Vector3(center);
@@ -49,19 +76,24 @@ export default class AxisAlignedBoundingBox {
     );
   }
 
-  transform(transformation) {
-    this.center.transformAsPoint(transformation);
-    // TODO - this.halfDiagonal.transformAsVector(transformation);
-    this.halfDiagonal.transform(transformation);
-    this.minimum.transform(transformation);
-    this.maximum.transform(transformation);
+  /**
+   * Applies a 4x4 affine transformation matrix to a bounding sphere.
+   * @param transform The transformation matrix to apply to the bounding sphere.
+   * @returns itself, i.e. the modified BoundingVolume.
+   */
+  transform(transform: readonly number[]): this {
+    this.center.transformAsPoint(transform);
+    // TODO - this.halfDiagonal.transformAsVector(transform);
+    this.halfDiagonal.transform(transform);
+    this.minimum.transform(transform);
+    this.maximum.transform(transform);
     return this;
   }
 
   /**
    * Determines which side of a plane a box is located.
    */
-  intersectPlane(plane) {
+  intersectPlane(plane: Plane): INTERSECTION {
     const {halfDiagonal} = this;
     const normal = scratchNormal.from(plane.normal);
     const e =
@@ -82,14 +114,13 @@ export default class AxisAlignedBoundingBox {
     return INTERSECTION.INTERSECTING;
   }
 
-  // Computes the estimated distance from the closest point on a bounding box to a point.
-  distanceTo(point) {
+  /** Computes the estimated distance from the closest point on a bounding box to a point. */
+  distanceTo(point: readonly number[]): number {
     return Math.sqrt(this.distanceSquaredTo(point));
   }
 
-  // Computes the estimated distance squared from the closest point on a bounding box to a point.
-  // A simplified version of OrientedBoundingBox.distanceSquaredTo
-  distanceSquaredTo(point) {
+  /** Computes the estimated distance squared from the closest point on a bounding box to a point. */
+  distanceSquaredTo(point: readonly number[]): number {
     const offset = scratchVector.from(point).subtract(this.center);
     const {halfDiagonal} = this;
