+++ modules/geospatial/src/constants.ts
