+++ modules/geospatial/src/ellipsoid/ellipsoid.ts
@@ -2,7 +2,16 @@
 // See LICENSE.md and https://github.com/AnalyticalGraphicsInc/cesium/blob/master/LICENSE.md
 
 /* eslint-disable */
-import {Vector3, Matrix4, toRadians, toDegrees, assert, equals, _MathUtils} from '@math.gl/core';
+import {
+  Vector3,
+  Matrix4,
+  toRadians,
+  toDegrees,
+  assert,
+  equals,
+  _MathUtils,
+  NumericArray
+} from '@math.gl/core';
 import * as vec3 from 'gl-matrix/vec3';
 
 import {WGS84_RADIUS_X, WGS84_RADIUS_Y, WGS84_RADIUS_Z} from '../constants';
@@ -20,17 +29,29 @@ const scratchCartesian = new Vector3();
 
 let wgs84;
 
-// A quadratic surface defined in Cartesian coordinates by the equation
-// <code>(x / a)^2 + (y / b)^2 + (z / c)^2 = 1</code>.  Primarily used
-// to represent the shape of planetary bodies.
+/**
+ * A quadratic surface defined in Cartesian coordinates by the equation
+ * `(x / a)^2 + (y / b)^2 + (z / c)^2 = 1`.  Primarily used
+ * to represent the shape of planetary bodies.
+ */
 export default class Ellipsoid {
-  // An Ellipsoid instance initialized to the WGS84 standard.
-  static get WGS84() {
-    wgs84 = wgs84 || new Ellipsoid(WGS84_RADIUS_X, WGS84_RADIUS_Y, WGS84_RADIUS_Z);
-    return wgs84;
-  }
+  /** An Ellipsoid instance initialized to the WGS84 standard. */
+  static readonly WGS84 = new Ellipsoid(WGS84_RADIUS_X, WGS84_RADIUS_Y, WGS84_RADIUS_Z);
 
-  // Creates an Ellipsoid from a Cartesian specifying the radii in x, y, and z directions.
+  readonly radii: Vector3;
+  readonly radiiSquared: Vector3;
+  readonly radiiToTheFourth: Vector3;
+  readonly oneOverRadii: Vector3;
+  readonly oneOverRadiiSquared: Vector3;
+  readonly minimumRadius: number;
+  readonly maximumRadius: number;
+  readonly centerToleranceSquared: number = _MathUtils.EPSILON1;
+  readonly squaredXOverSquaredZ: number;
+
+  /** Creates an Ellipsoid from a Cartesian specifying the radii in x, y, and z directions. */
+  constructor(x: number, y: number, z: number);
+  constructor();
+
   constructor(x = 0.0, y = 0.0, z = 0.0) {
     assert(x >= 0.0);
     assert(y >= 0.0);
@@ -58,8 +79,6 @@ export default class Ellipsoid {
 
     this.maximumRadius = Math.max(x, y, z);
 
-    this.centerToleranceSquared = _MathUtils.EPSILON1;
-
     if (this.radiiSquared.z !== 0) {
       this.squaredXOverSquaredZ = this.radiiSquared.x / this.radiiSquared.z;
     }
@@ -67,17 +86,20 @@ export default class Ellipsoid {
     Object.freeze(this);
   }
 
-  // Compares this Ellipsoid against the provided Ellipsoid componentwise and returns
-  equals(right) {
+  /** Compares this Ellipsoid against the provided Ellipsoid componentwise */
+  equals(right: Ellipsoid): boolean {
     return this === right || Boolean(right && this.radii.equals(right.radii));
   }
 
-  // Creates a string representing this Ellipsoid in the format '(radii.x, radii.y, radii.z)'.
-  toString() {
+  /** Creates a string representing this Ellipsoid in the format '(radii.x, radii.y, radii.z)'. */
+  toString(): string {
     return this.radii.toString();
   }
 
-  // Converts the provided cartographic to Cartesian representation.
+  /** Converts the provided cartographic to Cartesian representation. */
+  cartographicToCartesian(cartographic: number[], result: Vector3): Vector3;
+  cartographicToCartesian(cartographic: number[], result?: number[]): number[];
+
   cartographicToCartesian(cartographic, result = [0, 0, 0]) {
     const normal = scratchNormal;
     const k = scratchK;
@@ -96,8 +118,11 @@ export default class Ellipsoid {
     return k.to(result);
   }
 
-  // Converts the provided cartesian to cartographic (lng/lat/z) representation.
-  // The cartesian is undefined at the center of the ellipsoid.
+  /** Converts the provided cartesian to cartographic (lng/lat/z) representation.
+   * The cartesian is undefined at the center of the ellipsoid. */
+  cartesianToCartographic(cartesian: number[], result: Vector3): Vector3;
+  cartesianToCartographic(cartesian: number[], result?: number[]): number[];
+
   cartesianToCartographic(cartesian, result = [0, 0, 0]) {
     scratchCartesian.from(cartesian);
     const point = this.scaleToGeodeticSurface(scratchCartesian, scratchPosition);
@@ -118,11 +143,32 @@ export default class Ellipsoid {
     return toCartographicFromRadians([longitude, latitude, height], result);
   }
 
-  // Computes a 4x4 transformation matrix from a reference frame with an east-north-up axes
-  // centered at the provided origin to the provided ellipsoid's fixed reference frame.
+  /** Computes a 4x4 transformation matrix from a reference frame with an east-north-up axes
+   * centered at the provided origin to the provided ellipsoid's fixed reference frame. */
+  eastNorthUpToFixedFrame(origin: number[], result?: Matrix4): Matrix4;
+  eastNorthUpToFixedFrame(origin: number[], result: number[]): number[];
+
   eastNorthUpToFixedFrame(origin, result = new Matrix4()) {
     return localFrameToFixedFrame(this, 'east', 'north', 'up', origin, result);
   }
+
+  /** Computes a 4x4 transformation matrix from a reference frame centered at
+   * the provided origin to the ellipsoid's fixed reference frame.
+   */
+  localFrameToFixedFrame(
+    firstAxis: string,
+    secondAxis: string,
+    thirdAxis: string,
+    origin: Readonly<NumericArray>,
+    result?: Matrix4
+  ): Matrix4;
+  localFrameToFixedFrame<Matrix4T>(
+    firstAxis: string,
+    secondAxis: string,
+    thirdAxis: string,
+    origin: Readonly<NumericArray>,
+    result: number[]
+  ): number[];
 
   // Computes a 4x4 transformation matrix from a reference frame centered at
   // the provided origin to the ellipsoid's fixed reference frame.
@@ -130,13 +176,17 @@ export default class Ellipsoid {
     return localFrameToFixedFrame(this, firstAxis, secondAxis, thirdAxis, origin, result);
   }
 
-  // Computes the unit vector directed from the center of this ellipsoid toward
-  // the provided Cartesian position.
+  /** Computes the unit vector directed from the center of this ellipsoid toward
+   * the provided Cartesian position. */
+  geocentricSurfaceNormal(cartesian: number[], result?: number[]): number[];
+  geocentricSurfaceNormal<NumArray>(cartesian: number[], result: NumArray): NumArray;
   geocentricSurfaceNormal(cartesian, result = [0, 0, 0]) {
     return scratchVector.from(cartesian).normalize().to(result);
   }
 
-  // Computes the normal of the plane tangent to the surface of the ellipsoid at provided position.
+  /** Computes the normal of the plane tangent to the surface of the ellipsoid at provided position. */
+  geodeticSurfaceNormalCartographic<NumArray>(cartographic: number[], result: NumArray): NumArray;
+  geodeticSurfaceNormalCartographic(cartographic: number[]): number[];
   geodeticSurfaceNormalCartographic(cartographic, result = [0, 0, 0]) {
     const cartographicVectorRadians = fromCartographicToRadians(cartographic);
 
@@ -152,20 +202,24 @@ export default class Ellipsoid {
     return scratchVector.to(result);
   }
 
-  // Computes the normal of the plane tangent to the surface of the ellipsoid at the provided position.
+  /** Computes the normal of the plane tangent to the surface of the ellipsoid at the provided position. */
+  geodeticSurfaceNormal<NumArrayT>(cartesian: number[], result: NumArrayT): NumArrayT;
+  geodeticSurfaceNormal(cartesian: number[]): number[];
   geodeticSurfaceNormal(cartesian, result = [0, 0, 0]) {
     return scratchVector.from(cartesian).scale(this.oneOverRadiiSquared).normalize().to(result);
   }
 
-  // Scales the provided Cartesian position along the geodetic surface normal
-  // so that it is on the surface of this ellipsoid.  If the position is
-  // at the center of the ellipsoid, this function returns undefined.
+  /** Scales the provided Cartesian position along the geodetic surface normal
+   * so that it is on the surface of this ellipsoid.  If the position is
+   * at the center of the ellipsoid, this function returns undefined. */
+  scaleToGeodeticSurface(cartesian: number[], result?: number[]): number[];
   scaleToGeodeticSurface(cartesian, result) {
     return scaleToGeodeticSurface(cartesian, this, result);
   }
 
-  // Scales the provided Cartesian position along the geocentric surface normal
-  // so that it is on the surface of this ellipsoid.
+  /** Scales the provided Cartesian position along the geocentric surface normal
+   * so that it is on the surface of this ellipsoid. */
+  scaleToGeocentricSurface(cartesian: number[], result?: number[]): number[];
   scaleToGeocentricSurface(cartesian, result = [0, 0, 0]) {
     scratchPosition.from(cartesian);
 
@@ -185,19 +239,26 @@ export default class Ellipsoid {
     return scratchPosition.multiplyScalar(beta).to(result);
   }
 
-  // Transforms a Cartesian X, Y, Z position to the ellipsoid-scaled space by multiplying
-  // its components by the result of `Ellipsoid#oneOverRadii`
+  /** Transforms a Cartesian X, Y, Z position to the ellipsoid-scaled space by multiplying
+   * its components by the result of `Ellipsoid#oneOverRadii` */
+  transformPositionToScaledSpace(position: number[], result?: number[]): number[];
   transformPositionToScaledSpace(position, result = [0, 0, 0]) {
     return scratchPosition.from(position).scale(this.oneOverRadii).to(result);
   }
 
-  // Transforms a Cartesian X, Y, Z position from the ellipsoid-scaled space by multiplying
-  // its components by the result of `Ellipsoid#radii`.
+  /** Transforms a Cartesian X, Y, Z position from the ellipsoid-scaled space by multiplying
+   * its components by the result of `Ellipsoid#radii`. */
+  transformPositionFromScaledSpace(position: number[], result?: number[]): number[];
   transformPositionFromScaledSpace(position, result = [0, 0, 0]) {
     return scratchPosition.from(position).scale(this.radii).to(result);
   }
 
-  // Computes a point which is the intersection of the surface normal with the z-axis.
+  /** Computes a point which is the intersection of the surface normal with the z-axis. */
+  getSurfaceNormalIntersectionWithZAxis(
+    position: number[],
+    buffer?: number,
+    result?: number[]
+  ): number[];
   getSurfaceNormalIntersectionWithZAxis(position, buffer = 0.0, result = [0, 0, 0]) {
     // Ellipsoid must be an ellipsoid of revolution (radii.x == radii.y)
     assert(equals(this.radii.x, this.radii.y, _MathUtils.EPSILON15));
