+++ modules/geospatial/src/ellipsoid/helpers/scale-to-geodetic-surface.ts
